requirement.txt
boto3
discord.py
python-dotenv
cryptography>=42.0.0